+++ internal/handler/inbound_order/inbound_order.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/service"
+	service "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/service/inbound_order"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api/apperrors"
 	"github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/api/response"
 	models "github.com/varobledo_meli/W17-G10-Bootcamp.git/pkg/models/inbound_order"
